Advent of Code day 1
Started the challenge with a refresher of the AWK scripting language.
Splitting each group of inputs into a record and implicitly separating
fields with a newline feels clever but makes it difficult to see the
structure of the input while operating. The "shift" of first to second,
second to third, and so on is also probably inefficient. At least I'm
not sorting the input!