Advent of Code day 3
Yeah, it's looking like AWK all the way for me. I'm doing it. I can do
this. And I know, getline is ugly and awful, but without a good record
separator between the fields I can't do a ton. I could also save a ton
of time breaking out of loops and such, but this solution is decent
enough.