Advent of Code day 5
AWK's association lists are a hell of a thing. Indices are cast to
strings implicitly, two-dimensional "arrays" are not supported, only
NAME tokens and explicit lvalues are allowed as subscripts, and you
cannot remove elements once inserted. You can emulate stacks and queues
with a separate "array" like I have here, by passing the name of the
stack/queue (here I chose 1 through 9) as the first argument. Know that
stack[i,j] is a flat mapping of strings to other strings, and not nine
separate segments of memory. Still a neat solution though!