Advent of Code day 7
This solution feels a little hacky. Splitting the "path" constructed by
changing directories doesn't seem to work properly, nor does checking
the value of size["/"]. Ignoring the "$ cd /", "ls", and "dir [a-z]+"
lines was neat, as was avoiding the standard tree data structure with an
association list and a non-recursive walk back up the tree to add each
file's size to its ancestors. I also got a kick out of setting
PROCINFO["sorted_in"] to "@ind_str_asc" and printing the size of each
directory in a clear order with a for-each pattern (removed for
cleanliness).